Draw with view model

+++ Source/Simulation/SimulationViewModelGraph+Factory.swift
@@ -1,10 +1,10 @@
-import Foundation
+import UIKit
 
 extension SimulationViewModelGraph {
     static func factoryWithEntropy(entropy:Entropy) -> SimulationViewModelGraph {
         var graph:SimulationViewModelGraph = SimulationViewModelGraph()
-        graph.chaos = entropy.chaos
-        graph.productivity = entropy.productivity
+        graph.chaos = CGFloat(entropy.chaos)
+        graph.productivity = CGFloat(entropy.productivity)
         return graph
     }
 }
